+++ Mods ChoGGi/RC Constructor Routes/metadata.lua
@@ -1,0 +1,16 @@
+return PlaceObj("ModDef", {
+	"title", "RC Constructor Routes v0.1",
+	"version", 1,
+	"saved", 1545307200,
+	"image", "Preview.png",
+	"id", "ChoGGi_RCConstructorRoutes",
+	"steam_id", "1595958450",
+	"author", "ChoGGi",
+	"lua_revision", LuaRevision,
+	"code", {
+		"Code/Script.lua",
+	},
+	"description", [[For some reason it intentionally isn't supposed to be able to create routes...
+
+Requested by veryinky.]],
+})